requirements.txt
functions-framework==3.*
google-cloud-aiplatform==1.22.1
google-cloud-appengine-logging==1.1.1
google-cloud-audit-log==0.2.0
google-cloud-bigquery
google-cloud-bigquery-storage==2.13.1
google-cloud-bigtable==2.9.0
google-cloud-core==2.3.0
google-cloud-dataproc==4.0.2
google-cloud-datastore==2.6.0
google-cloud-firestore==2.5.0
google-cloud-kms==2.11.1
google-cloud-language==2.4.2
google-cloud-logging==3.1.1
google-cloud-monitoring==2.9.1
google-cloud-pubsub==1.7.0
google-cloud-resource-manager==1.5.0
google-cloud-scheduler==2.6.3
google-cloud-spanner==3.14.0
google-cloud-speech==2.14.0
google-cloud-storage==2.3.0
google-cloud-tasks==2.9.0
google-cloud-translate==3.7.3
google-cloud-videointelligence==2.7.0
google-cloud-vision==2.7.2
google-resumable-media==2.3.3
pandas==1.3.5
Werkzeug==2.2.2
Flask==2.2.3
pytz==2023.3
fsspec==2023.5.0
gcsfs==2023.5.0
pandas-gbq==0.23.0
prophet==1.1.5